moved push Logik into push service js
