+++ backend/server.ts
@@ -25,6 +25,8 @@ declare global {
       readonly BACKEND_PORT: string
       readonly MONGO_URL: string
       readonly BACKEND_SAVE_REPORTS_ON_DISK: 'TRUE' | 'FALSE'
+      readonly VITE_PUBLIC_VAPID_KEY: string
+      readonly VITE_PRIVATE_VAPID_KEY: string
       readonly VITE_MAX_FILE_SIZE: string
       readonly RATE_LIMIT_WINDOW_MS: string
       readonly RATE_LIMIT: string
